+++ src/app/admin/pages/list-words/list-words.component.ts
@@ -1,7 +1,9 @@
 import { ComponentType } from '@angular/cdk/portal';
-import { Component, OnInit } from '@angular/core';
+import { ChangeDetectorRef, Component, OnInit } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
+import { finalize, tap } from 'rxjs';
 import { ResponseWordDTO } from 'src/app/models/response/responseWordDTO';
+import { ErrorMessageService } from 'src/app/services/error-message.service';
 import { FormWordsComponent } from '../../components/form-words/form-words.component';
 import { ConnectionService } from '../../services/connection.service';
 
@@ -14,11 +16,14 @@ import { ConnectionService } from '../../services/connection.service';
 export class ListWordsComponent implements OnInit {
   public listWords: ResponseWordDTO[] = []
   public listDeleteElement: any[] = []
-  public load: boolean = true
+  public load: boolean = false;
   public componentForm: ComponentType<FormWordsComponent>;
   public dialogRef!: MatDialogRef<FormWordsComponent, any>
 
-  constructor(private connectionS: ConnectionService, public dialog: MatDialog) {
+  constructor(private connectionS: ConnectionService,
+    public dialog: MatDialog,
+    private cdRef: ChangeDetectorRef,
+    private errorMsgS: ErrorMessageService) {
     this.componentForm = FormWordsComponent
   }
 
@@ -28,8 +33,23 @@ export class ListWordsComponent implements OnInit {
 
   getListWords() {
     this.connectionS.getWords()
+      .pipe(tap({
+        next: (res) => {
+          this.load = true
+          this.listWords = res
+          this.cdRef.markForCheck()
+        },
+        error: (err) => {
+          this.load = true
+          this.cdRef.markForCheck()
+        }
+      }),
+        finalize(() => {
+          setTimeout(() => {
+            this.load = false
+          }, 300)
+        }))
       .subscribe((resp) => {
-        this.listWords = resp
       })
   }
 
@@ -39,7 +59,7 @@ export class ListWordsComponent implements OnInit {
       height: '600px',
       data: {},
       disableClose: true,
-      panelClass:'custom-dialog-container'
+      panelClass: 'custom-dialog-container'
     });
     this.dialogRef.afterClosed().subscribe(result => {
       if (result) {
@@ -51,11 +71,15 @@ export class ListWordsComponent implements OnInit {
 
   deleteWord() {
     if (this.listDeleteElement.length > 0 && this.listWords.length > 0) {
-      this.listDeleteElement.forEach(id => {
-        this.connectionS.deleteWords(id).subscribe(res => {
-          this.procesaPropagar()
-        })
-      });
+      this.errorMsgS.deleteElement().then((res) => {
+        if (res.isConfirmed) {
+          this.listDeleteElement.forEach(id => {
+            this.connectionS.deleteWords(id).subscribe(res => {
+              this.procesaPropagar()
+            })
+          });
+        }
+      })
     }
   }
   procesaPropagar() {
